+++ Executor/Tasks/PreRun/coordinate.py
@@ -7,7 +7,7 @@ from Interfaces import RemoteApi
 
 class Coordinate(Task):
     def __init__(self, logMethod, parent):
-        super().__init__("Coordinate", parent, logMethod, None)
+        super().__init__("Coordinate", parent, None, logMethod, None)
 
     def Run(self):
         remote = self.parent.Descriptor.Remote
@@ -18,7 +18,6 @@ class Coordinate(Task):
                 if host is not None:
                     remoteApi = RemoteApi(host, port)
                     self.parent.RemoteApi = remoteApi
-                    # TODO: Why are these messages not visible in the logs?
                     self.Log(Level.INFO, 'Remote connection configured. Waiting for remote Execution ID...')
 
                     timeout = eastWest.Timeout or 120