+++ Executor/Tasks/Run/publish_from_source.py
@@ -9,8 +9,10 @@ class PublishFromSource(Task):
         super().__init__(name, parent, params, logMethod, None)
         self.paramRules = {
             'Pattern': (None, True),
-            'Keys': (None, True),
-            'Path': (None, False)  # Mandatory only for PublishFromFile, handled below
+            'Keys': ([], False),
+            'Path': (None, False),  # Mandatory only for PublishFromFile, handled below
+            'VerdictOnMatch': ("NotSet", False),
+            'VerdictOnNoMatch': ("NotSet", False),
         }
 
     def Run(self):
@@ -19,6 +21,8 @@ class PublishFromSource(Task):
         filePath = self.params["Path"]
         pattern = self.params["Pattern"]
         keys = self.params["Keys"]
+        onMatch = self.GetVerdictFromName(self.params["VerdictOnMatch"])
+        onNoMatch = self.GetVerdictFromName(self.params["VerdictOnNoMatch"])
 
         self.Log(Level.DEBUG, f"Looking for pattern: '{pattern}'; Assigning groups as:")
 
@@ -26,21 +30,26 @@ class PublishFromSource(Task):
             for index, key in keys:
                 self.Log(Level.DEBUG, f"  {index}: {key}")
         except Exception as e:
+            self.SetVerdictOnError()
             raise RuntimeError(f"Invalid 'Keys' definition: {e}")
 
         regex = re.compile(pattern)
 
+        matchFound = False
         for line in self.generator({"Path": filePath}):
             match = regex.match(line)
             if match:
                 self.Log(Level.INFO, f"Match found: {match.string}")
+                matchFound = True
                 for index, key in keys:
                     self.Publish(key, match.group(index))
+        self.Verdict = onMatch if matchFound else onNoMatch
 
     def generator(self, params: Dict):
         raise NotImplementedError()
 
     def raiseConfigError(self, variable: str):
+        self.SetVerdictOnError()
         raise RuntimeError(f"'{variable}' not defined, please review the Task configuration.")
 
 
@@ -49,7 +58,7 @@ class PublishFromPreviousTaskLog(PublishFromSource):
         super().__init__("Publish From Previous Task Log", parent, params, logMethod)
 
     def generator(self, params: Dict):
-        logMessages = self.parent.Params["PreviousTaskLog"]
+        logMessages = self.parent.PreviousTaskLog
         for message in logMessages:
             yield message