+++ Executor/Tasks/Run/slice_creation_time.py
@@ -34,6 +34,7 @@ class SliceCreationTime(Task):
                 nestData = json.load(input)
         except Exception as e:
             self.Log(Level.ERROR, f"Exception while reading NEST file: {e}")
+            self.SetVerdictOnError()
             return
 
         from Helper import InfluxDb, InfluxPayload, InfluxPoint  # Delayed to avoid cyclic imports
@@ -51,6 +52,7 @@ class SliceCreationTime(Task):
                 sliceId = response
             except Exception as e:
                 self.Log(Level.ERROR, f"Exception on instantiation, skipping iteration: {e}")
+                self.SetVerdictOnError()
                 sleep(pollTime)
                 continue
 
@@ -86,6 +88,7 @@ class SliceCreationTime(Task):
                                  f"Deployment time for slice {sliceId} (Iteration {iteration}): {ns_depl_time}")
                     except Exception as e:
                         self.Log(Level.ERROR, f"Exception while calculating deployment time, skipping iteration: {e}")
+                        self.SetVerdictOnError()
                         break
 
                     point = InfluxPoint(datetime.now(timezone.utc))
@@ -113,6 +116,7 @@ class SliceCreationTime(Task):
                         self.Log(Level.DEBUG, f"Waiting for slice deletion.")
             except Exception as e:
                 self.Log(Level.ERROR, f"Exception while deleting slice: {e}")
+                self.SetVerdictOnError()
 
         self.Log(Level.DEBUG, f"Payload: {payload}")
         self.Log(Level.INFO, f"Sending results to InfluxDb")
@@ -120,6 +124,7 @@ class SliceCreationTime(Task):
             InfluxDb.Send(payload)
         except Exception as e:
             self.Log(Level.ERROR, f"Exception while sending payload: {e}")
+            self.SetVerdictOnError()
             if csvFile is None:
                 self.Log(Level.INFO, "Forcing creation of CSV file")
                 csvFile = join(self.parent.TempFolder, f"SliceCreationTime.csv")