+++ Facility/facility.py
@@ -1,12 +1,12 @@
-from os.path import abspath, join
-import yaml
+from os.path import abspath
 from .action_information import ActionInformation
 from .dashboard_panel import DashboardPanel
 from .resource import Resource
 from Helper import Log, Level
-from typing import Dict, List, Tuple, Callable, Optional
+from typing import Dict, List, Tuple, Optional
 from threading import Lock
 from Utils import synchronized
+from .Loader import Loader, ResourceLoader, ScenarioLoader, UeLoader, TestCaseLoader
 
 
 class Facility:
@@ -23,7 +23,7 @@ class Facility:
     ues: Dict[str, List[ActionInformation]] = {}
     testCases: Dict[str, List[ActionInformation]] = {}
     extra: Dict[str, Dict[str, object]] = {}
-    dashboards: Dict[str, List[ActionInformation]] = {}
+    dashboards: Dict[str, List[DashboardPanel]] = {}
     resources: Dict[str, Resource] = {}
     scenarios: Dict[str, Dict] = {}
 
@@ -31,193 +31,49 @@ class Facility:
 
     @classmethod
     def Reload(cls):
-        from Helper import IO
-        allParameters: Dict[str, Tuple[str, str]] = {}
-
-        def _ensureFolder(path: str):
-            if not IO.EnsureFolder(path):
-                cls.Validation.append((Level.INFO, f'Auto-generated folder: {path}'))
-
-        def _loadFolder(path: str, kind: str, callable: Callable):
-            ignored = []
-            for file in IO.ListFiles(path):
-                if file.endswith('.yml'):
-                    cls.Validation.append((Level.INFO, f'Loading {kind}: {file}'))
-                    callable(join(path, file))
-                else:
-                    ignored.append(file)
-            if len(ignored) != 0:
-                cls.Validation.append((Level.WARNING,
-                                       f'Ignored the following files on the {kind}s folder: {(", ".join(ignored))}'))
-
-        def _loadFile(path: str) -> Optional[Dict]:
-            try:
-                with open(path, 'r', encoding='utf-8') as file:
-                    raw = yaml.safe_load(file)
-                    return raw
-            except Exception as e:
-                cls.Validation.append((Level.ERROR, f"Unable to load file '{path}': {e}"))
-                return None
-
-        def _get_ActionList(data: Dict) -> List[ActionInformation]:
-            actionList = []
-            for action in data:
-                actionInfo = ActionInformation.FromMapping(action)
-                if actionInfo is not None:
-                    actionList.append(actionInfo)
-                else:
-                    cls.Validation.append((Level.ERROR, f'Action not correctly defined for element (data="{action}").'))
-                    actionList.append(ActionInformation.MessageAction(
-                        'ERROR', f'Incorrect Action (data="{action}")'
-                    ))
-            if len(actionList) == 0:
-                cls.Validation.append((Level.WARNING, 'No actions defined'))
-            else:
-                for action in actionList:
-                    cls.Validation.append((Level.DEBUG, str(action)))
-            return actionList
-
-        def _get_PanelList(data: Dict) -> List[DashboardPanel]:
-            panelList = []
-            for panel in data:
-                try:
-                    parsedPanel = DashboardPanel(panel)
-                    valid, error = parsedPanel.Validate()
-                    if not valid:
-                        cls.Validation.append((Level.ERROR, f'Could not validate panel (data={panel}) - {error}'))
-                    else:
-                        panelList.append(parsedPanel)
-                except Exception as e:
-                    cls.Validation.append((Level.ERROR,
-                                           f"Unable to parse Dashboard Panel (data={panel}), ignored. {e}"))
-            cls.Validation.append((Level.DEBUG, f'Defined {len(panelList)} dashboard panels'))
-            return panelList
-
-        def _testcaseLoader(path: str):
-            try:
-                data = _loadFile(path)
-
-                allKeys = list(data.keys())
-                dashboard = data.pop('Dashboard', None)
-                standard = data.pop('Standard', None)
-                custom = data.pop('Custom', None)
-                distributed = data.pop('Distributed', False)
-                parameters = data.pop('Parameters', {})
-
-                if dashboard is None:
-                    cls.Validation.append((Level.WARNING, f'Dashboard not defined. Keys: {allKeys}'))
-
-                if standard is None:
-                    standard = (custom is None)
-                    cls.Validation.append((Level.WARNING,
-                                           f'Standard not defined, assuming {standard}. Keys: {allKeys}'))
-                keys = list(data.keys())
-
-                if len(keys) > 1:
-                    cls.Validation.append((Level.ERROR, f'Multiple TestCases defined on a single file: {list(keys)}'))
-                for key in keys:
-                    testCases[key] = _get_ActionList(data[key])
-
-                    extra[key] = {
-                        'Standard': standard,
-                        'PublicCustom': (custom is not None and len(custom) == 0),
-                        'PrivateCustom': custom if custom is not None else [],
-                        'Parameters': parameters,
-                        'Distributed': distributed
-                    }
-
-                    if dashboard is not None:
-                        dashboards[key] = _get_PanelList(dashboard)
-
-                for name, info in parameters.items():
-                    type, desc = (info['Type'], info['Description'])
-                    if name not in allParameters.keys():
-                        allParameters[name] = (type, desc)
-                    else:
-                        oldType, oldDesc = allParameters[name]
-                        if type != oldType or desc != oldDesc:
-                            cls.Validation.append(
-                                (Level.WARNING, f"Redefined parameter '{name}' with different settings: "
-                                                f"'{oldType}' - '{type}'; '{oldDesc}' - '{desc}'. "
-                                                f"Cannot guarantee consistency."))
-            except Exception as e:
-                cls.Validation.append((Level.ERROR, f'Exception loading TestCase file {path}: {e}'))
-
-        def _ueLoader(path: str):
-            try:
-                data = _loadFile(path)
-                keys = data.keys()
-                if len(keys) > 1:
-                    cls.Validation.append((Level.WARNING, f'Multiple UEs defined on a single file: {list(keys)}'))
-                for key in keys:
-                    if key in ues.keys():
-                        cls.Validation.append((Level.WARNING, f'Redefining UE {key}'))
-                    actions = _get_ActionList(data[key])
-                    ues[key] = actions
-            except Exception as e:
-                cls.Validation.append((Level.ERROR, f'Exception loading UE file {path}: {e}'))
-
-        def _resourceLoader(path: str):
-            try:
-                data = _loadFile(path)
-                resource = Resource(data)
-                if resource.Id in resources.keys():
-                    cls.Validation.append((Level.WARNING, f'Redefining Resource {resource.Id}'))
-                resources[resource.Id] = resource
-            except Exception as e:
-                cls.Validation.append((Level.ERROR, f'Exception loading Resource file {path}: {e}'))
-
-        def _scenarioLoader(path: str):
-            try:
-                data = _loadFile(path)
-                if len(data.keys()) > 1:
-                    cls.Validation.append((Level.WARNING, f'Multiple Scenarios defined on a single file: {list(keys)}'))
-                for key, value in data.items():
-                    if key in scenarios.keys():
-                        cls.Validation.append((Level.WARNING, f'Redefining Scenario {key}'))
-                    scenarios[key] = value
-                    cls.Validation.append((Level.DEBUG, f'{key}: {value}'))
-            except Exception as e:
-                cls.Validation.append((Level.ERROR, f'Exception loading Resource file {path}: {e}'))
-
         cls.Validation.clear()
 
-        # Generate all folders
+        # Generate missing folders
         for folder in [cls.TESTCASE_FOLDER, cls.UE_FOLDER, cls.RESOURCE_FOLDER, cls.SCENARIO_FOLDER]:
-            _ensureFolder(folder)
+            v = Loader.EnsureFolder(folder)
+            cls.Validation.extend(v)
 
-        testCases = {}
-        ues = {}
-        dashboards = {}
-        extra = {}
-        scenarios = {}
-
+        resources = cls.resources
         if len(cls.BusyResources()) != 0:
-            resources = cls.resources
             cls.Validation.append((Level.WARNING, "Resources in use, skipping reload"))
         else:
-            resources = {}
-            _loadFolder(cls.RESOURCE_FOLDER, "Resource", _resourceLoader)
+            ResourceLoader.Clear()
+            v = ResourceLoader.LoadFolder(cls.RESOURCE_FOLDER, "Resource")
+            cls.Validation.extend(v)
+            resources = ResourceLoader.GetCurrentResources()
 
-        _loadFolder(cls.TESTCASE_FOLDER, "TestCase", _testcaseLoader)
-        _loadFolder(cls.UE_FOLDER, "UE", _ueLoader)
-        _loadFolder(cls.SCENARIO_FOLDER, "Scenario", _scenarioLoader)
+        TestCaseLoader.Clear()
+        v = TestCaseLoader.LoadFolder(cls.TESTCASE_FOLDER, "TestCase")
+        cls.Validation.extend(v)
 
-        for collection, name in [(testCases, "TestCases"), (ues, "UEs"),
-                                 (dashboards, "DashBoards"), (resources, "Resources"),
-                                 (scenarios, "Scenarios")]:
+        UeLoader.Clear()
+        v = UeLoader.LoadFolder(cls.UE_FOLDER, "UE")
+        cls.Validation.extend(v)
+
+        ScenarioLoader.Clear()
+        v = ScenarioLoader.LoadFolder(cls.SCENARIO_FOLDER, "Scenario")
+        cls.Validation.extend(v)
+
+        cls.resources = resources
+        cls.testCases = TestCaseLoader.GetCurrentTestCases()
+        cls.extra = TestCaseLoader.GetCurrentTestCaseExtras()
+        cls.dashboards = TestCaseLoader.GetCurrentDashboards()
+        cls.ues = UeLoader.GetCurrentUEs()
+        cls.scenarios = ScenarioLoader.GetCurrentScenarios()
+
+        for collection, name in [(cls.testCases, "TestCases"), (cls.ues, "UEs"),
+                                 (cls.dashboards, "DashBoards"), (cls.resources, "Resources"),
+                                 (cls.scenarios, "Scenarios")]:
             keys = collection.keys()
             if len(keys) == 0:
                 cls.Validation.append((Level.WARNING, f'No {name} defined on the facility.'))
             else:
                 cls.Validation.append((Level.INFO, f'{len(keys)} {name} defined on the facility: {(", ".join(keys))}.'))
-
-        cls.ues = ues
-        cls.testCases = testCases
-        cls.extra = extra
-        cls.dashboards = dashboards
-        cls.resources = resources
-        cls.scenarios = scenarios
 
     @classmethod
     def GetUEActions(cls, id: str) -> List[ActionInformation]: